Moved the state of the API fetch to a parent component
